refactor: mypage-파일구조 변경
