+++ src/app/mypage/components/header/index.ts
@@ -1,0 +1,4 @@
+export { MypageHeader } from './MypageHeader';
+export { MainTabSection } from './MainTabSection';
+export { SubTabSection } from './SubTabSection';
+export { ReviewToggleButton } from './ReviewToggleButton';
