+++ src/app/mypage/components/index.ts
@@ -1,4 +1,4 @@
-export { TabSection, type MainTabType, type SubTabType } from './TabSection';
+export * from './header';
 export { MeetingDividingCard } from './Card/MeetingDividingCard';
 export { MeetingShoppingCard } from './Card/MeetingShoppingCard';
 export { ProfileSideBar } from './layout/ProfileSideBar';
