+++ src/components/Molecules/EmptyState/EmptyState.tsx
@@ -1,7 +1,10 @@
 'use client';
 
 import { Button } from '@/components';
-import { MainTabType, SubTabType } from '@/app/mypage/components';
+import {
+  type MainTabType,
+  type SubTabType,
+} from '@/app/mypage/utils/mypageType';
 import { EMPTY_STATE_CONFIG } from './constants';
 import Image from 'next/image';
